requirements.txt
groq
python_dotenv
ffmpeg
pyaudio
speechrecognition
pydub
elevenlabs
gradio